requirements.txt
aws-cdk-lib==2.208.0
constructs>=10.0.0,<11.0.0
python-dotenv>=1.0.0
boto3>=1.34.0
urllib3>=1.26.0
typing-extensions>=4.0.0
cdk-nag>=2.37.0
